+++ bignbit/get_collection_concept_id.py
@@ -37,7 +37,7 @@ class CMA(Process):
         # Use override for collection concept id from dataset config if provided
         collection_id = dataset_config.get('concept_id')
         if collection_id is None:
-          collection_id = get_collection_concept_id(collection_shortname, cmr_provider, cmr_environment)
+            collection_id = get_collection_concept_id(collection_shortname, cmr_provider, cmr_environment)
         self.input['collection_concept_id'] = collection_id
         return self.input